% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_function.R
\name{arrange_function}
\alias{arrange_function}
\title{Arranging the data by where the game was played and in descending order}
\usage{
arrange_function(basketball_data)
}
\arguments{
\item{basketball_data}{the data frame we are using to do the manipulation}
}
\description{
change to desc(Venue) to arrange in descending order
}
