% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_2022_games.R
\name{func_2022_games}
\alias{func_2022_games}
\title{Filtering all games played in 2022}
\usage{
func_2022_games(basketball_data)
}
\arguments{
\item{basketball_data}{the data frame we are using to do the manipulation}
}
\value{
return the data set with only 2022 games
}
\description{
Filtering all games played in 2022
}
