% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_function.R
\name{mutate_function}
\alias{mutate_function}
\title{Create a new column that designates where each game was
played. Also create a column that gives the score differences (team1−team2)}
\usage{
mutate_function(basketball_data)
}
\arguments{
\item{basketball_data}{the data frame we are using to do the manipulation}
}
\description{
Mutate where the game was played into a new column
We are just going to use the second name column to do this
}
\details{
The second piece of the code is to get the difference in the scores
The absolute value function is used to not get any negative scores
}
