% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_function.R
\name{read_function}
\alias{read_function}
\title{Function to read in the data}
\usage{
read_function(fileUrl)
}
\arguments{
\item{fileUrl}{the file we are reading from}
}
\value{
The function will return basketball data as a tibble
}
\description{
Function to read in the data
}
