% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_function.R
\name{remove_function}
\alias{remove_function}
\title{Removing the 6th and 7th columns from the basketball data set}
\usage{
remove_function(basketball_data)
}
\arguments{
\item{basketball_data}{the data frame we are using to do the manipulation}
}
\description{
As of now the columns have not been named so just using V6 and V7
}
