% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_basketball_app.R
\name{shiny_basketball_app}
\alias{shiny_basketball_app}
\title{Shiny Basketball Package function}
\usage{
shiny_basketball_app()
}
\description{
This function will create the shiny app for the basketball package.
The appdir object basically looks for the shiny folder in the basketball
package and runs the ui and server functions.
}
\details{
The ui.R file will create the ui for the application.
The server.R will set up the server.
The app.R function sources both the ui and server functions.
}
