% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_function.R
\name{summarize_function}
\alias{summarize_function}
\title{Function that will summarize the data for a team win loss percentage record}
\usage{
summarize_function(basketball_data, team_name)
}
\arguments{
\item{basketball_data}{the data frame we are using to do the manipulation}
}
\description{
Function that will summarize the data for a team win loss percentage record
}
