% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_function.R
\name{team_function}
\alias{team_function}
\title{Function that will filter the tibble to only games played by a given team}
\usage{
team_function(basketball_data, team_name)
}
\arguments{
\item{basketball_data}{the data frame we are using to do the manipulation}

\item{team_name}{the name of the team of whose games we want}
}
\description{
Function that will filter the tibble to only games played by a given team
}
